Added persistent storage to the clock servers.
Adding persistent storage to the clock servers complicates the design quite a lot. The "root" of the complication is that writing to disk is considered an async operation, and therefore cannot be done on the main executor thread. If we were to reconsider and change it so that writing to persistent storage is considered synchronous, and can be done in the middle of the code, then things would get a lot easier. Could be worth revisiting this decision.
Also, changed from tabs to spaces which screws up the change history.